Update Java version of sock shop.
Summary:
We update the Java version used by the sock shop demo to (greatly) improve rendered flamegraphs. We believe this is because of increased support for preserving frame pointers (and secondarily, because the Java binary includes symbols).

To do this, we needed to self host the sock shop images on our own gcr.io path.

Test Plan: Opened up the sock shop in a browswer and verified as much functionality as possible. Nothing regressed.

Reviewers: #stirling, michelle

Reviewed By: michelle

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11896

GitOrigin-RevId: ebbab72b8c27f5e1df17b1dec849358f53d4c0aa