Perf profiler: for symbolization, use BCC library functions directly.
Summary: This diff changes the way we wrap BCC symbolization. Previously, we used a BPF stack traces table, which included some amount of symbolization support (although we had to extend its support to meet our needs). In this diff, we switch over to directly using the BPF library functions which gives us more flexibility (and is congruent to the way the BPF python API uses the BPF library).

Test Plan: Existing

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Subscribers: yzhao

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11909

GitOrigin-RevId: 7991cd4fd0f895eac236660e50818731fe4846ef