Add summary views for each of the timeseries functions in pxviews
Summary:
Added pod_network_summary,container_process_summary, inbound_http_throughput_summary to the pxviews library. The timeseries
versions are not simple to export because it has many windows. A single
window is easy to export. Each of these are basically the
single window version of each timeseries script.

Test Plan: Ran each script and ensured the results are as expected

Reviewers: jamesbartlett, michelle, vihang

Reviewed By: vihang

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11989

GitOrigin-RevId: 39a502d91be2b917fd41c3016c3ccaa099a32ff1