gRPC-c probes
Summary:
1 Initiate PerCPU variables of the gRPC-c eBPF module.
2 Look for new processes that use the gRPC-c library. Determine the library's version by its MD5 hash.
Currently, only 4 library hashes have been added. In the future, we will need to either develop a mechanism that automatically finds hashes, or determine the library's version dynamically.
3 When a process with a gRPC-c library is found, attach 6 needed probes.

Test Plan: Bpf test to follow in separate diff. This code is feature flagged such that it is not on by default.

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Signed-off-by: Ori Shussman <ori@groundcover.com>

Differential Revision: https://phab.corp.pixielabs.ai/D12087

GitOrigin-RevId: b8c7e92c0b7c46f1603444990bdd92f68ecf5463