#550: Honor `postMessage` above URL param for theme
Summary:
When Pixie is embedded, there are a few ways to set its theme both at load time and later. Currently, the former "wins" if the latter is done too soon.

This makes sure the most recent attempt to set the theme is the one that actually applies.

Fixes #550

Test Plan: Embed Pixie with the `theme=light` URL parameter. As soon as it's ready, send a `postMessage({ pixieStyles: ... })` message. Without this fix, the `postMessage` gets quietly clobbered because the URL parameter was loaded too late. With this fix, the `postMessage` gets priority even if it's a little early.

Reviewers: michelle, vihang

Reviewed By: michelle

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D11992

GitOrigin-RevId: ec15091747389a3a33602477a6238f6f36494fa9