Expose compiler_state inside Dataframe qlobject
Summary:
I am adding a feature that requires a compiler_state pointer inside of Dataframe: parsing
string times inside of the DataFrame instead of through a rule.
This diff does the api rewriting, while the next one will actually use the compiler_state.

Test Plan:
Ran tests and still works. Just an addition of a field to the structure so shouldn't
change anything.

Reviewers: michelle, jamesbartlett, vihang

Reviewed By: jamesbartlett

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12015

GitOrigin-RevId: e98ba111997560584ad291b649b12152864b3d31