Fix bazel wrapper script to ignore other repos.
Summary:
Fixes the bazel wrapper script to not add config values if the repo's workspace name is not `px`.
Before this diff, running bazel in a different repo would fail because it doesn't have a config named clang.

Test Plan: Tested on `rules_go` and the pixie repo and both worked with just a simple `bazel` command.

Reviewers: zasgar, vihang

Reviewed By: vihang

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12012

GitOrigin-RevId: 04dfb5b502d63279514aaa7ab05c114575e3197e