Remove aggregated state use by vizier monitor
Summary:
This diff removes aggregated state usage by the vizierMonitor and instead relies on individual vizier states of pvc/node to update status.

The purpose of removing the aggregated state is to eventually allow for a feature to repair a failed pvcState. When pvcState fails, a feature to repair it by switching to etcd can be triggered as a fix.

`getVizierState()` is used to get updates on pvcState/nodeState by checking the aggregatedState. In the case where a failed pvcState is fixed by switching to etcd, the aggregatedState holds on to the failed pvcState and wont get updated. This causes the repair feature to be called numerous times even when the cluster is healthy and backed by etcd because of an inactive failed pvcState. Having a check to see when PV/etcd is used instead of relying on an aggregatedState and then returning with the appropriate failed state can be beneficial to avoid triggering a repair function in the wrong scenario.

Test Plan: Ran test: `bazel test //src/operator/controllers:controllers_test --test_output=streamed` and tested with skaffold operator

Reviewers: philkuz, vihang

Reviewed By: vihang

Signed-off-by: Kartik Pattaswamy <kpattaswamy@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12003

GitOrigin-RevId: cf96c911b7193f387d2137647f40dd5367b60aad