Upgrade finagle to version that includes tls tracing compatible netty-tcnative
Summary: This PR upgrades finagle to install a version that includes the netty changes I upstreamed (netty/netty#12438, netty/netty-tcnative#731) to support #407. Once this is upgraded, the next step is to refresh #463.

Test Plan: Ran the thriftmux container and verified that the logs of the client show that netty tcnative is used (improperly installed netty tcnative will silently succeed)

Reviewers: #stirling, jps

Reviewed By: #stirling, jps

Signed-off-by: Dom Del Nano <ddelnano@twitter.com>

Differential Revision: https://phab.corp.pixielabs.ai/D12111

GitOrigin-RevId: d3da5e233e5aa38b1b29865fd4cc8c58128966c2