Implement changes to get finagle tls tracing working.
Summary:
This PR will complete the initial support of #407 and is a refresh of an earlier proof of concept change (#463).

The goal is to provide an extensible way for supporting other boringssl based tracing targets if the target binary supports the following:

 -  Ability to determine the target binary's openssl version through the RawFptrManager or through dlsym
    The boringssl implementation populate's openssl's bio_st data structure with the fd of the socket file descriptor (like I did with netty -

 -  Add SSL.bioSetFd method to JNI code to support Pixie integration netty/netty-tcnative#731).

This won't work for envoy's case, but the hope is that this is more similar to Pixie's dynamic openssl tracing while envoy is more similar to nodejs (requiring additional application specific integration).

Closes #569.

Test Plan:
New test verifies that TLS tracing is successful (output below)
Verified the changes to socket_trace_connector allow `TraceMode::OnNewerKernel` to work

Reviewers: #stirling, zasgar

Reviewed By: zasgar

Signed-off-by: Dom Del Nano <ddelnano@twitter.com>

Differential Revision: https://phab.corp.pixielabs.ai/D12324

GitOrigin-RevId: 5ce5cd67c32ae9d6e4b11e04ab854d65371a3a16