[perf_tool 1/n] Add skeleton for coordinator service.
Summary: This diff adds a bunch of boilerplate for the coordinator service. The coordinator service, will be in charge of receiving requests to run experiments, launching the necessary preparation tasks, and then telling the runner service to run the experiment.

Test Plan: N/A. No functional code in this diff just boilerplate/creds etc.

Reviewers: vihang, zasgar, michelle

Reviewed By: zasgar

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12455

GitOrigin-RevId: 582743a68dbdec7b69a4469e84d154c26223b682