[perf_tool 9/n] Implement datastore for coordinator service.
Summary:
Implements the experiment queue datastore for the coordinator service.
The datastore stores each experiment's spec along with how many prerequisite tasks still need to be completed before the experiment can be run by the runner.

Test Plan: Adds a test for the datastore, that exercises all the sql queries.

Reviewers: vihang, zasgar, michelle

Reviewed By: zasgar

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12465

GitOrigin-RevId: b73f44ca2481fc6f6750894c3c7b99b6a0d5d017