Fix bug where indexer would only flush when new events came in
Summary:
Indexer would only flush events to elastic when the batch size > maxActions or when an event
came in after some certain timeout. This missed the case where items come in before the timeout period.

This diff creates a flusher routine. Whenever a new event shows up, the routine schedules a flush. If the maxBatch size is reached,
we call flush immediately. But if it's not reached, a timer will trigger the flush at the end of the period.

Test Plan: Added some tests and deploy to plc-dev

Reviewers: nserrino, michelle, vihang

Reviewed By: vihang

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12106

GitOrigin-RevId: ff17db9be77b125aea3494d5cc7290f81d0432ca