Updating a plugin should move old preset scripts to custom scripts
Summary:
The previous behavior was that updating a plugin will result in any old, deleted preset scripts to be deleted.
However, it may be the case that users will want to retain any old preset scripts. This change updates the plugin update process so that any old preset scripts are just moved to custom scripts.

Test Plan: Unit test

Reviewers: vihang, philkuz

Reviewed By: philkuz

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12151

GitOrigin-RevId: ab53bb046955f9c259d867f35996602eb44fc962