[GH-582] Create GetUnusedVarName() that parses a PxL script and returns a var name that has not been used in the script.
Summary:
This diff implements GetUnusedVarName() a utility which returns a var name that has not been used in the script. Users pass in a desired var name
and the function will add suffixes to the var name until the paired script does not already contain a variable with that name.

Will be used inside of GenerateOTelScript to ensure that we can assign the exported dfs to an unused variable.

Test Plan: Added unit tests

Reviewers: michelle, vihang, jamesbartlett, nserrino

Reviewed By: jamesbartlett

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12167

GitOrigin-RevId: f1aaf1f818188d1fa3cd9256845a7f84702a0808