[GH-582] Setup GenerateOTelScript through the passthrough proxy
Summary:
In this diff I wired up GenerateOTelScript through the Passthrough proxy.
The method is similar to ExecuteScript except instead of being a server side
stream, we're just doing simple request/response. I ended up re-using the same machinery as before
and for the cloud side, implementing the service interface for the GenerateOTelScript.

This is the last step to make GenerateOTelScript available to Pixie clients that access the backend.

Note: I also added a new error message that is returned whenever Vizier receives a message that it doesn't handle.
I found that it just hung before which was a bad experience.

Test Plan: Access GenerateOTelScript through the go api and was able to generate a script.

Reviewers: michelle, vihang, jamesbartlett, nserrino

Reviewed By: michelle, jamesbartlett

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12172

GitOrigin-RevId: ff355c474a1a2357a9ede65b74166dab3453033f