Upgrade google/benchmark to v1.7.0.
Summary:
We upgrade our dependency on google/benchmark to v1.7.0. This is to bring in the feature that allows us to name a specific benchmark, i.e. as follows:
`BENCHMARK(BM_std_map)->Name("std::map/out-of-cache")->Args(out_of_cache_args);`
The out of cache args may be an inscrutable set of numbers contrived to show the desired behavior. When running the benchmark without the `Name()` being set, one will see `BM_std_map/65536/16` or similar, i.e. an inscrutable name. When name is used, we see instead a name that captures the intent and meaning of this set of benchmark args.

Test Plan: Existing.

Reviewers: #stirling, #third_party_approvers, yzhao

Reviewed By: #stirling, #third_party_approvers, yzhao

Subscribers: yzhao

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12173

GitOrigin-RevId: 3645c6a436df5769e86a2d21fd1ca32b6b5f93c6