[GH-582] Add Export to Plugin button to the editor for ScratchPad scripts
Summary:
Adding the export to Plugin button only for Scratch Pad scripts for now.
If users find it useful and demand it, we'll extend this for the other scripts,
but to steel-thread this we'll limit it to scratch scripts.
{F194652}

The following diffs will connect it to actually work with the rest of the system.

Test Plan: Tested that clicking the button works and renders.

Reviewers: nlanam, michelle, nserrino

Reviewed By: nlanam

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12193

GitOrigin-RevId: c805ebdbe5dc8ca2de1e9eb76e9d9fdd5c4d7197