[GH-582]: Wire up the Create Plugin Button to the GRPC Call
Summary:
This diff wires up the Create Plugin Script button to the proper GRPC call.
A few points for discussion
1. Had to expose the PxLEditorText directly. Ideally would be able to call "saveEditor()" but that currently causes the page to execute, which is not desirable IMO.
2. I write all the errors from this path into the resultsContext. I think that the view for errors isn't as good as it can be for users we populate the same area as
the tables, but that area is obscured by the editor. Most developers iteratively write their scripts, testing as they go. Having to collapse the editor to view the errors
adds an unnecessary barrier for their development.

3. I put the navigation logic inside of script_context. I'm not sure that I really like this approach, but I want to follow the pattern set by `execute` so did it this way.
I didn't feel very good to create a dep on EditorContext, so decided to allow a parameter, the pxlScript, to be passed into this function.

Test Plan:
Tested the following
1. Compiler error still renders
{F194562}
2. Error raised by OTelGenerate rendes
{F194554}
3. Valid OTel script is passed into the create page
{F194564}

Reviewers: nlanam, michelle, nserrino

Reviewed By: nlanam

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12195

GitOrigin-RevId: c55d8eb34f68c226b43cc38574aa09070705fad3