[GH-582] Disable Generate button and show spinner while waiting for results
Summary:
I found that running the Generate OTel script against a cluster that was not yet supported was confusing. You would click the button then wait for something to happen. Sometimes nothing returned.

This diff addresses the problem by disabling the button while the script is running and shows a spinner while we're waiting.

I think the spinner is mainly useful while not all clusters support GenerateOTelScript, so I think we should delete it when those scripts are gone.

Test Plan: Tested on dev UI

Reviewers: nlanam, nserrino

Reviewed By: nlanam

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12207

GitOrigin-RevId: 267a7b268cacd3221ee224202b4f82389dbc2a41