GenerateOTelExport polish: Zero data columns should throw an error
Summary:
Discovered that zero data columns causes a problem right now. I think that it should
throw an error so that users make sure they actually add something.

In my opinion it's better to throw an error and ask the user to remedy it than returning the empty
list.

Test Plan: Added a unit test and tried on skaffold'd vizier.

Reviewers: michelle, vihang, nserrino

Reviewed By: michelle

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12208

GitOrigin-RevId: 87bf7ca9cdd447df341b64549179be0d6fa84fc1