Implementing AstToString for Return and Assign types
Summary:
In a follow-up diff I need to generate string reprs for Return and Assign AST types
in OTel Generation, so I added support here.

Made the signature accept the base AST type and simplified the test execution as a result.

Test Plan: added unit tests for return and assign.

Reviewers: vihang, michelle, nserrino

Reviewed By: nserrino

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12225

GitOrigin-RevId: 8903a02fefa8b5f7b6c8e49dd4ebc50794c2dc5b