[GH-576] Plugin job should perform auto-updates
Summary:
We want to ensure that it is easy to keep plugins up-to-date. Currently, users have to manually enable/reenable plugins to update them.
Instead, as soon as we load up new plugins in the load_db job, we can auto-update any out-of-date plugins.
This change updates the load_db job so that after new plugins are loaded from github, it also runs through all of the existing plugins.
Any existing plugins with a new version will be updated.

Closes #576

Test Plan:
Deployed to a dev + testing cloud. Tested updating plugins with:
- Adding/removing configs
- Existing custom scripts: scripts should remain functional
- Removing preset scripts: script becomes a new custom script
- Creating new preset scripts: new preset script is added

Reviewers: vihang, philkuz

Reviewed By: philkuz

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12200

GitOrigin-RevId: 83a5f077d30587066c281e8514dcfaca40ccfddf