Implement JetStream streamer
Summary:
This implements the streamer interface backed by JetStream. This is
the intreface we use for STAN and this implementation should make it relatively
simple to swap out STAN for JetStream.

Note: Consumer management is done inside the interface. We basically only
have two types of consume interfaces. The Peek interface uses an ephemeral
consumer. The PersistentSubscribe creates a single consumer for the durable
and binds to the same consumer on all calls. This is to ensure that the
consumer isn't automatically deleted by the library thus helping us keep a
cursor into the stream.

Test Plan: Added unit tests to test the behavior.

Reviewers: michelle, philkuz

Reviewed By: michelle

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12249

GitOrigin-RevId: 7bb78ec2c91b366bc6613c07c4ca3cf5bb6ad747