[GH-593] Command Palette: Parser
Summary:
The heart of the Command Palette when writing script commands.

Takes the input text, the selection (caret), and yeilds the following information:
- What key/value pairs it contains (`foo:bar baz:"a quoted string"` -> `{ foo: 'bar', baz: 'a quoted string' }`)
- A list of tokens, classified so that they can be syntax-highlighted in presentation elsewhere.
- Links related tokens together, so the `['foo', ':', 'bar']` tokens can directly refer to one another.
- Which tokens are touched or covered by the selection (caret), and where within them it covers.
- Basic error detection, such as unclosed quotes.

This function caches its results, so that consumers can just call it with the same inputs in different places (that might be hard to connect otherwise) and not have to run the parse again.

It will be consumed both by the input's syntax highlighting overlay, and the PxL command provider.

Test Plan: `yarn test`. Consumers will be in other diffs.

Reviewers: michelle, vihang, philkuz

Reviewed By: michelle

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12263

GitOrigin-RevId: 4181640489cdd4f866a5390f2013a803c9cd4fcf