Refactor ObjectDeleter to use provided kubernetes Clientset.
Summary: I was trying to reuse the ObjectDeleter outside of the CLI, and I noticed that it doesn't actually use the provided Clientset for most of its operations. In git history, its clear that it used to use the provided Clientset but that was removed because it was really slow. This diff adds that back but with the Burst and QPS of the config set to the kubectl defaults, so deletes are just as fast as before.

Test Plan: Tested `px delete` on a 2 node cluster. Before this change it took 28-29s and after this change it still took 28-29s.

Reviewers: michelle, vihang

Reviewed By: michelle

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12290

GitOrigin-RevId: d8ec6551fef8bd3df0ece3f395020ae0f3c2446a