Update operator to provide OperatorVersion in vizier status.
Summary: We want to track the operator version in the vizier heartbeat. D12242 added the OperatorVersion field to the CRD. This diff updates the operator to provide the OperatorVersion value when a vizier monitor is created.

Test Plan:
Skaffold the operator
Check the vizier status with `kubectl get vizier -n pl -o json`
Observe new `operatorVersion` field is populated

Reviewers: vihang, philkuz, michelle

Reviewed By: michelle

Signed-off-by: Hannah Troisi <htroisi@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12243

GitOrigin-RevId: dea95249a732ace341167a15d1e332777a024729