Handle custom image registries in Vizier templates
Summary:
We are adding support for custom image registries for Vizier. This is done by templatizing our Vizier YAMLs to use the custom image registry or default image registry based on whether a registry has been specified.
We expect the image paths to be the custom image registry prepended to the default image path, with "/" substituted with "-" (since many registries like docker do not support nested paths)

Test Plan: Run template generator and inspect output of the templates.

Reviewers: vihang, zasgar, jamesbartlett

Reviewed By: vihang

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12299

GitOrigin-RevId: 9cef3ee08a236aafaa394a94744f70676ad82ab8