Rename cloudconn clusterrole
Summary:
This clusterrole was named as a part of cleanup. This didn't lead to any issues when trying to test an update (by building an rc), since the existing clusterrole doesn't get deleted/the old rolebinding continues to exist.
However, clusterrolebindings don't expect to be updated, so this leads to errors when running `skaffold deploy`.

Test Plan: Skaffold deploy

Reviewers: vihang, jps

Reviewed By: jps

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12304

GitOrigin-RevId: e3deb25dac512105d2e20a4764873e5a23a17e32