Remove byteman from thriftmux container
Summary: This was a workaround that was relied on when I was testing an earlier version of #569, but it is no longer required. See the details on #407 for more background on this issue.

Test Plan: Verified that mux trace bpf test passes -- verifies that thriftmux docker image builds after these changes

Reviewers: #stirling, rcheng

Reviewed By: #stirling, rcheng

Signed-off-by: Dom Del Nano <ddelnano@twitter.com>

Differential Revision: https://phab.corp.pixielabs.ai/D12318

GitOrigin-RevId: 99770d28e8041102d0279d2e13d0de18cb5d0a26