Cleanup dev_dns_updater for MacOS
Summary:
The `SUDO_USER` based homedir computation was incorrect when running on
MacOS. Instead let's use `os.HomeDir` to determine the dir, and parse flags
before we elevate to sudo. This should now be easier to use cross platform.

Test Plan:
Ran it on macOS without specifying the kubeconfig location. Ensured
that it worked as expected.

Reviewers: zasgar, michelle, philkuz

Reviewed By: michelle

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12329

GitOrigin-RevId: 2d7b5872ab23cf9aa47900b48ba793978eb1caa8