Add mux_data pxl script for later use in finagle demo
Summary:
This adds a mux_data pxl script that shows mux traffic in a table. This will be used in #607 so that users can have a way to visualize the traffic deployed from the future finagle demo. The pxl script was developed as part of Twitter's hackathon in the finagle-helloworld repo to demonstrate that the mux parser was functional when it was developed earlier this year.

Closes #612

Test Plan: Verified that src/e2e_test/vizier/planner:planner_test test succeeds after adding the script and fails if the mux UDF is misspelled (indicating that the mux_data pxl script is being leveraged during the test)

Reviewers: #stirling, htroisi

Reviewed By: htroisi

Subscribers: michelle

Signed-off-by: Dom Del Nano <ddelnano@twitter.com>

Differential Revision: https://phab.corp.pixielabs.ai/D12346

GitOrigin-RevId: 0e2726c775bf498fdf10ca31447a81cdcdb32ad6