[GH-601] Add buttons to zoom and and out of request graphs
Summary:
They appear when the request graph is focused or hovered. Clicking them does what you would expect.

Fixes #601

Test Plan:
Click on or hover over a request graph. A pair of buttons should appear in the bottom right to zoom in/out; this should work the same as using the mouse wheel does.

Note: the focus behavior when doing this and then trying to interact with the graph directly is a little fussy; you may need to click an extra time. This is an existing bug with all of the buttons in that corner.

Reviewers: michelle, vihang

Reviewed By: vihang

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12372

GitOrigin-RevId: 51ac712440209fd27974c4f60fe953d0d9647f18