Update OIDC client params
Summary:
**Prompt for login**
Pixie's auth logout doesn't propagate a logout to the underlying OIDC
provider. This means that logging out and logging in automatically drops you
back into the same account (assuming that the user is still logged in with the
underlying provider.) This makes account switching more challenging.

**Skip prompt for refetch**
When refetching the token (on an auth callback), ensure to not prompt and use
the existing token.

**Skip loading userInfo**
The UserManager loads UserInfo by default but we don't use it directly. Disable
userinfo load to speed up login.

Test Plan: Tested OIDC auth and Auth0 auth in a test cluster.

Reviewers: michelle, philkuz, zasgar

Reviewed By: michelle

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12405

GitOrigin-RevId: bb4a17f28b53e109dc0ef345e80dfd582cf43d89