Fix gRPC cpp deps in pl_cc_proto_library to ensure proper linking
Summary:
In another diff, linker failed to link grpc_codegen_init.cc file to a generated grpc library.
The bug was strange because it only appeared in some random test, while other deps of the proto
linked fine.

Turns out that we had a dep on grpc++ in other parts of the code so we never saw this error.

This diff explicitly adds the necessary dep to the `pl_cc_proto_library`.
There's some risk that in grpc ~1.50 this will change significantly and break our deps again.

We alternatively considered linking to `grpc++`, but we felt like that would drastically increase
the size of our test objects + increase our compilation time of tests so we decided against that
for now.

Test Plan: Test the broken test with this change and the test started to succeed

Reviewers: vihang, jamesbartlett

Reviewed By: vihang

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12416

GitOrigin-RevId: cddbc44c8485e0c06c54a9ff507b9abba07f1cd8