[update bcc/bpftrace 1/3] Add libbpf as separate dependency.
Summary:
Previously, we used libbpf that was submodule'd into bcc. However, new versions of bpftrace are moving away from using bcc (although they still require bcc) and instead trying to use libbpf directly.
This means that libbpf is now a direct dependency of bpftrace, so we add libbpf as an external repo, so that we can point both bcc and bpftrace to the same version of libbpf

This diff just adds libbpf as a repo and doesn't use it anywhere.

Test Plan: This diff just adds a repo, no functional changes.

Reviewers: zasgar, vihang, oazizi, #third_party_approvers

Reviewed By: oazizi, #third_party_approvers

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12429

GitOrigin-RevId: 6235f1c0ebba6bc19789f06834462da4b85ee3ff