Cleanup plugin load image builds
Summary:
This was in a halfway state because we switched when/how these get
built and run. Now that the job image is built as part of `cloud_images_push`
we no longer need the manual builds.

Test Plan:
skaffold dev cloud builds and runs the latest load_db image.
Prod cloud releases should build, push and apply the appropriate image.

Reviewers: michelle, philkuz

Reviewed By: michelle

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12464

GitOrigin-RevId: 8b569b04c48535b4b3fd3bb58613bffc31a1d511