Fix UI build caching
Summary:
The UI build deps included all of `.yarn`, however `.yarn/cache` is
large and gitigored. We should infact only depend on the stuff that's checked
into the repo.
So followed https://next.yarnpkg.com/getting-started/qa#which-files-should-be-gitignored
to update ignores and gitattributes.
Update the build deps so that the UI build hopefully becomes a lot less volatile.

Test Plan: bazel build for the ui still works.

Reviewers: michelle, zasgar

Reviewed By: michelle

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12469

GitOrigin-RevId: 6dc314dd73c3c95c579544bad38fc098d6aa5ae3