Don't strip Redis error prefix
Summary:
Currently the Redis protocol strips the single character prefix for each type of data. This is mostly fine for most data - however, it makes it impossible for downstream services cannot determine whether a response is a failure or not.
This diff adds the error prefix back in.

github: #645

Test Plan: Updated tests

Reviewers: #stirling, philkuz

Reviewed By: philkuz

Signed-off-by: Ryan Cheng <rcheng@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12532

GitOrigin-RevId: 7b3a501ee2f52fbbc39d95e925e0d156b74e240b