Add more informative error messaging in operator
Summary:
We recently added sentry to our operator, which has resulted in some spammy errors.

For errors where the vizier status fails to update due to a timeout, this usually means the K8s API has become overloaded and cannot respond in time. Updated this to have more informative messaging.
We have another error when the Vizier deploy fails, but its unclear at which stage it is failing. Updated the errors to be more fine-grained.

Test Plan: Skaffold ran operator

Reviewers: htroisi, philkuz

Reviewed By: htroisi

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12534

GitOrigin-RevId: ddb199537f8ecac2f294927adce1e96a8a098853