Update kafka demo zookeeper image tag
Summary: We recently updated the yaml to point to the `latest` tag for zookeeper. This can lead to problems if the owner made another push to the `latest` tag with an incompatible zookeeper version. Instead, I pushed the zookeeper image to a registry that we control and updated the image tag.

Test Plan: n/a

Reviewers: htroisi, rcheng, philkuz

Reviewed By: htroisi

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12535

GitOrigin-RevId: 15eeb34ab2a2bece87df1d8830804ef92997f34b