Remove Spinner from Create Export Script
Summary:
We had a spinner on CreateExport Script button that was there to better indicate that
something was happening if a cluster didn't have an export script endpoint. Now that it's past our
support timeline (2 mos) we can remove this feature.

Test Plan: Tested against a dev cloud and the export script feature still works.

Reviewers: nlanam, michelle, nserrino

Reviewed By: nserrino

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12540

GitOrigin-RevId: 8ab408c62c341b2b44c338c07c8a7d1a86d85ac2