QueryBroker: Move SyncScript() into goroutine
Summary:
SyncScripts() makes a call over the NatsBridge to get scripts. It blocks the starting
of the server, but takes forever to complete under certain circumstances. Those circumstances
still need to be investigated as I'm pretty sure it's a sign of a bug, but at least we should
make sure that clusters can start up in the first place.

Test Plan:
Had a cluster that failed to start because of the waittime on SyncScripts(). Ran this
version and saw that it successfully ran. Also tested against a cluster that had no such problems
and the cluster was still operational.

Reviewers: vihang, michelle, jamesbartlett, nserrino

Reviewed By: michelle

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12583

GitOrigin-RevId: 4f4df00987f869f3fc3fb826c7f6f21015aad7b3