[build] Fix mac build.
Summary: The GraalVM diff broke the mac build, because `local_java_repository` requires the `java_home` location to exist even if the rule is unused. This creates an empty directory on macos for the graalvm `java_home`.

Test Plan: Tested that the directory gets created. Relying on vihang to test on mac.

Reviewers: vihang, zasgar, michelle

Reviewed By: vihang

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12612

GitOrigin-RevId: f03dbf722e1bf2aa14f2e38a7ca53640557dbe20