Remove pl-cluster-config dependency from Metadata deployments
Summary:
`pl-cluster-config` was filling in the values for whether etcd was used for the Metadata
and for the path for the etcd service. We don't need to do this anymore because we have separate
yamls for the two different types.

I ran into a problem with this config based on how our operator works right now. When a user
switches their Vizier from etcdOPerator to not etcdOperator the change wouldn't fully flow
through because we wouldn't change the configuration. I looked at ways to make the configuration
work, but this ended up being the simplest.

Test Plan:
Tried to switch between the etcd metadata and persistent metadata without having
to change the pl-cluster-config configmap, which was not the case before.

Reviewers: vihang, michelle

Reviewed By: michelle

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12621

GitOrigin-RevId: aa911c5c9c27ad96ba61e520a4ff45cc2ebbe8d1