[GH-593] Command Palette: Icons
Summary:
Final change to release the preview of the Command Palette.

Changes the icons for suggestions to ones that make sense.

This is far from everything to make a rich command palette, but the foundation is all here now and it works well enough to be useful. Therefore, this will close the original open issue.

Fixes #593

Test Plan: Use the command palette to complete pods, namespaces, nodes, and services. All four of them should have sensible icons for their suggestions now instead of things like clotheshangers.

Reviewers: michelle, vihang

Reviewed By: michelle

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12630

GitOrigin-RevId: 0f72d9782bada68a4d503b0bd1eceab5badfe14f