Fix coverage build with new bazel
Summary:
The bazel 6 upgrade broke coverage builds. This was mainly due to
the bazel-out and bazel-testlogs symlinks getting cleared away. Instead
let's fix the UI covereage to handle the SF munging within the rule and tell
bazel to combine reports. This should reduce the complexity and also fix the
build.

Test Plan:
Enabled coverage for this phab run. Ensure that jenkins can run
it successfully to completion.

Reviewers: zasgar, michelle, jamesbartlett

Reviewed By: zasgar

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12656

GitOrigin-RevId: 7c02f46cd49d83cdbf20cfba58e582b51705a8f1