Unlink Jenkins fetch of source and target from builds
Summary:
The Jenkins helpers to fetch the source and target stashes were tied
to the pxbuild container in the default pod. However the BPF runs don't need
a pod with the pxbuild image or the resources requested by it.
So unlink the two and change the bpf test pods to just run a thin gcloud
container.

Test Plan:
The BPF stages should continue to work. The other tests should also
continue to work on Jenkins.

Reviewers: zasgar, michelle, jamesbartlett

Reviewed By: jamesbartlett

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12658

GitOrigin-RevId: 4d63f0f458d23eb7f0eaa1c86ae64f6c3bc3ae02