Add global.cluster recognition for clusterName
Currently, when this chart is installed as a subchart of the `nri-bundle` chart (where it is named `pixie-chart`), the `pixie-chart.clusterName` value must still be set, although `nri-bundle` already sets a required value `global.cluster` to store the cluster name. 

As a person installing the chart bundle with New Relic and Pixie, this is not a great experience as I must enter the same value twice - only for this value.

This change makes it unnecessary to set `pixie-chart.clusterName` for the chart bundle, as long as `global.cluster` is set.

Signed-off-by: Max Lemieux <mlemieux@newrelic.com>