Update internal LBs to be internal on AKS
Summary:
Adds the annotation specified in Azure's AKS docs,
https://learn.microsoft.com/en-us/azure/aks/internal-lb#create-an-internal-load-balancer,
so that the load balancers for the vzconn and cloud-proxy services
are assigned internal IP addresses. This matches the behavior of the GKE
load balancer annotation.

Test Plan:
Verify internal IP addresses for the vzconn and cloud-proxy services
when creating a new self-hosted Pixie on AKS.

Signed-off-by: Jeff Belgum <jeffbelgum@gmail.com>