Add autopilot versions to Vizier YAML templator
Summary:
A previous diff added an autopilot-specific version of Pixie to our YAMLs. This diff adds the autopilot versions to our template generator, so that they are packaged up in our `vizier_template_yamls.tar` during the release build.
The operator will be responsible filling in these templates depending on whether the user has set the `autopilot` flag or not.

Test Plan: Ran Vizier rc build

Reviewers: vihang, jamesbartlett

Reviewed By: jamesbartlett

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12663

GitOrigin-RevId: 521e5b0391125d06ca6df0349bdcc126f2fa0dd4