Shorten instance names to fit within 63 chars
Summary:
Some job names are long enough that we were trying to create instances
with names that were too long. This shortens them.
A more robust solution would be to detect the name length and uniquely shorten
them but this is a fine stopgap and will work for our current jobs.

Test Plan: Ran the script with long buildtags.

Reviewers: zasgar, michelle, jamesbartlett

Reviewed By: jamesbartlett

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12669

GitOrigin-RevId: 816d3bd4d8f5da9aebb75c444c406d88774301ed