Patch llvm cmake files for llvm libs.
Summary: When using llvm as a dependency for other cmake builds, cmake requires the llvm cmake configs. However, those configs check for all of the llvm libs being present. This patches the cmake configs to prevent checking for the libs, since we've removed some of them to save space.

Test Plan: Test as part of future diff.

Reviewers: zasgar, vihang, #third_party_approvers

Reviewed By: zasgar, #third_party_approvers

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12725

GitOrigin-RevId: c02d796dfb1a8f03e10f49c7ed84089db5c1d809