Add invite boolean as extra query param to OIDC provider
Summary:
The OIDC provider may want to know certain information about the signup, such as whether or not it was an invite, without having to read the state (it may not be able to).
This change adds an extra boolean to the params passed to the oidc provider to indicate whether or not the signup/login was initiated by following an invite link.

Test Plan: Deploy to a testing cloud and check the redirect link to the OIDC provider

Reviewers: vihang

Reviewed By: vihang

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12726

GitOrigin-RevId: 65f445487312455de8103008922d93c668236dd2