Drop reliance on gsutil for builds
Summary:
Having gsutil for a couple of public gcs files and also a gsutil
packaged but google-cloud-sdk was a bit problematic when we needed to access
files under auth. Switch the repo rules to be `http_file` since they are
public and have http access. Drop the tool so that we only have one gsutil.

Test Plan:
ran packer. jenkins builds should successfully upload bpf test results
to GCS now.

Reviewers: zasgar, jamesbartlett

Reviewed By: zasgar

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12760

GitOrigin-RevId: 69be01c383b6a4032deb6cf9b26a87a00d9c5dcc