[build] Fix bpftrace using clang includes from /opt.
Summary:
Bpftrace should use clang includes from the downloaded llvm libraries instead of from /opt.
This diff adds Clang_ROOT to tell cmake to use the right includes. It also requires patching the llvm lib distributions so that cmake doesn't check that all the libraries exist.

Test Plan: Tested with a patch that added a print out of where cmake is getting clang includes from. Made sure that it uses the downloaded ones and not /opt.

Reviewers: vihang, jamesbartlett, #third_party_approvers

Reviewed By: vihang, #third_party_approvers

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12792

GitOrigin-RevId: c14c59440aed1256bfcc9d267b1162710acb88af