Always detect runnable state of command properly
Summary:
There was an issue where the "Run" button in the command palette would sometimes change to a loading state when moving the caret around.

This fixes the issue, and also adds a default suggestion when the current input is a valid script command (that way, hitting Enter again will just run it as-is).

Test Plan: Try moving caret around in command palette while command is valid. As long as it remains valid, the Run button should stay blue and there should be a default suggestion that does the same thing as the Run button.

Reviewers: michelle, vihang

Reviewed By: michelle

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12813

GitOrigin-RevId: 09ef680d28fa6acbf8ad34bcd1c0fc839a495670