Fix runfiles resolution for cgroup_path_resolver_test.
Summary:
cgroup_path_resolver_test was failing when run standalone, as opposed to when run in bazel's linux sandbox.
The fundamental issue here is the tests assumption that runfiles under a common directory prefix (in our tree) will be under the same directory prefix in the runfiles manifest.
This assumption is core to the way this test works, so its hard to remove. Instead I made all of the runfiles it references exist in our tree. This way the assumption holds true.

Test Plan: Ran in standalone and also in the sandbox and both worked.

Reviewers: oazizi, zasgar, michelle, vihang

Reviewed By: vihang

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12815

GitOrigin-RevId: ab57c5955ef3ea80e28f2d8f396124ed7810fb3d