Collapse kustomize edit into main kustomize build rule
Summary:
We might need to do different edits to the same kustomization depending
on what we are building. But since the edited kustomization path has to depend
on the input path, we end up with conflicting bazel rules that output different
files to the same path. This is messy to resolve so just fold it all into the
main build command.

Test Plan: Ran the kustomize targets

Reviewers: jamesbartlett, zasgar

Reviewed By: jamesbartlett

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12879

GitOrigin-RevId: 0fd883cd96c4b2478cdf6600a045dddcbf33b532