Switch from bazel define flags to skylib flags
Summary:
We used --define to set the image versions and we explicitly created
targets for every version of the image push we needed.
Instead switch to skylib string_flags with TemplateVariableInfo and
BuildSettingInfo to parse the command line flags, with reasonable defaults and
use those to provide ctx vars that will be Make variable substituted by the
container_bundle rules.

Test Plan:
Build the container_push scripts with various flag settings and checked
the generated shell scripts. The image locations were as expected.
Also built the yamls to ensure that the kustomize replacements are working as
intended.

Reviewers: jamesbartlett, zasgar, michelle

Reviewed By: jamesbartlett

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12883

GitOrigin-RevId: 7dad1d2d696d1ed603e68f7b14048c0a7bbc9efa