Support templated replacements in kustomize_build
Summary:
This changes kustomize_build's handling of replacement and uses
the string format syntax combined with ctx.vars to enable Make var like
substitution.
This allows us to reduce the dupe in vizier images and just use the var instead.

Test Plan:
Built the vizier images with diffrent values for build_type.
Checked to ensure that the prefix and version were as expected.

Reviewers: jamesbartlett, zasgar, michelle

Reviewed By: jamesbartlett

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12887

GitOrigin-RevId: 261a91f30361c7e8a27dbdf8c8dbf7d76d2236c4