Remove unused and deprecated cluster_name field
Summary:
According to the apimachinery package:

> ClusterName is a legacy field that was always cleared by
> the system and never used

This field is renamed in `v0.24` and will be removed in `v0.25`.
We don't rely it on anywhere in our codebase and just pass it through during
object conversions and for tests. So let's remove it.

Test Plan: All existing tests continue to pass.

Reviewers: michelle, zasgar

Reviewed By: michelle

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12950

GitOrigin-RevId: d3ad6e414c1a92c54fac65d55b9f1fdb96af5128