Add configs for OSS Buildbuddy
Summary: We want buildbuddy for our OSS builds. This diff adds the configs that are needed to deploy buildbuddy, along with the jenkinsfile changes to connect to the OSS buildbuddy for our OSS builds.

Test Plan: Test by triggering builds on jenkins.px.dev, ensuring that build results end up on buildbuddy and are viewable without being logged in.

Reviewers: vihang, zasgar, jamesbartlett

Reviewed By: vihang

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12945

GitOrigin-RevId: bb5930969038c605fb3e472516dd339e8a1a59b5