[build/sysroots] Update sysroot creator to parse debian package list directly.
Summary:
This removes the usage of apt from the sysroot creator. Instead we parse the package list directly.
As part of this change, we also now support an exclude list to remove depedencies.

Test Plan: Tested with arm and x86 as part of unreleased code. No functional changes to existing build.

Reviewers: zasgar, vihang, #third_party_approvers

Reviewed By: vihang, #third_party_approvers

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12944

GitOrigin-RevId: 5d92beb9d50012a249e9892199617ab16440606d