Hide GH_API_KEY from buildbuddy.
Summary:
Buildbuddy hides the environment variables by default, but not values set through action_env=KEY=VALUE.
This diff changes the jenkins.bazelrc to use client_env instead of action_env.
This is not currently and issue since buildbuddy is private. But we are going to make a public build buddy soon, so this will become an issue otherwise.

Test Plan: Tested that client_env doesn't show up in buildbuddy. Relying on existing tests to make sure the GH_API_KEY is still passed properly.

Reviewers: zasgar, michelle

Reviewed By: michelle

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12981

GitOrigin-RevId: 6f6f6326e58c63896b107198e190638f2722d113