Add operator release build to jcasc and fix buildURL (#746)
Summary: This change adds the operator release build in jcasc, which
includes adding the necessary plugins.

This change also fixes the jenkins build URL. Originally, we were adding
the build URL in the jcasc configs. The config would be successfully
loaded when the Jenkins pod first starts up, since the build URL isn't
populated at that point. However, if we were to reload the config with
the Jenkins pod already running, the config would fail to load properly
because we will try to rewrite the buildURL which is now populated in
the "defaultConfig". Turns out we can properly specify the build URL as
a part of the ingress.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Deploy to Jenkins

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>