Fix PR description linter to ignore code blocks and details.
Summary: PR description linter should ignore code blocks and details, when running its checks.
This allows for example for me to put the following details block that has lines that begin with whitespace.
<details>
      Starts with whitespace.
                ends with whitespace
</details>

Type of change: /kind cleanup

Test Plan: Testing with this PR, also tested locally with a variety of combinations of details, code blocks.

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>